+++ src/main/java/com/givememoney/controller/GrantController.java
@@ -4,7 +4,7 @@ import com.givememoney.dto.common.ResponseDto;
 import com.givememoney.dto.government.request.GovernmentServiceListRequestDto;
 import com.givememoney.dto.government.response.ServiceDetailListResDto;
 import com.givememoney.dto.government.response.ServiceListResDto;
-import com.givememoney.service.SubsidyService;
+import com.givememoney.service.GrantService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,15 +14,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/v1")
 @RequiredArgsConstructor
-public class SubsidyController {
+public class GrantController {
 
-    private final SubsidyService subsidyService;
+    private final GrantService grantService;
 
     @GetMapping("/gov24/v1/serviceList")
     public ResponseDto getServiceList(int page, int perPage) {
 
         GovernmentServiceListRequestDto requestDto = new GovernmentServiceListRequestDto(page, perPage);
-        ServiceListResDto responseDto = subsidyService.getServiceList(requestDto);
+        ServiceListResDto responseDto = grantService.getServiceList(requestDto);
         return ResponseDto.builder()
                 .messsage("success")
                 .data(responseDto)
@@ -32,7 +32,7 @@ public class SubsidyController {
 
     @GetMapping("/gov24/v1/serviceDetail")
     public ResponseDto getServiceDetail(String serviceId) {
-        ServiceDetailListResDto responseDto = subsidyService.getServiceDetails(serviceId);
+        ServiceDetailListResDto responseDto = grantService.getServiceDetails(serviceId);
         return ResponseDto.builder()
                 .messsage("success")
                 .data(responseDto)
